Relocates leadership API client to common and removes client-specific initialisation.
